// src/app/shared/models/company.ts
export class Company {
  sellerCompanyId: number;
  password: string;
  sellerCompanyBin: string;
  sellerCompanyEmail: string;
  sellerCompanyMobilePhone: string;
  sellerCompanyNameEn: string;
  sellerCompanyNameKk: string;
  sellerCompanyNameRu: string;
  sellerCompanyPhone: string;
  sellerCompanyName: {
    en: string;
    kk: string;
    ru: string;
  };
  username: string;
}